Added 64bit support to DLL. MQL5 raw support added.

+++ ThriftMQL/FXMindMQLClient.cs
@@ -14,7 +14,7 @@ namespace FXBusinessLogic.BusinessObjects.Thrift
     public class FXMindMQLClient : ThriftClient<FXMindMQL.Client> 
     {
 
-        public FXMindMQLClient(ushort port)
+        public FXMindMQLClient(long port)
         {
             try
             {
